faust_producer/requirements.txt
faust-streaming
pandas
motor
